Visual Studio 2005: Fixed type conversion warnings.

+++ tests/src/generator_blas3.cpp
@@ -60,7 +60,7 @@ using namespace boost::numeric;
 //
 static const unsigned int min_large_block_size = 32;
 static const unsigned int max_large_block_size = 128;
-static const unsigned int n_large_blocks = static_cast<unsigned int>(std::log(max_large_block_size/min_large_block_size)/std::log(2)+1.0);
+static const unsigned int n_large_blocks = static_cast<unsigned int>(std::log(static_cast<double>(max_large_block_size/min_large_block_size))/std::log(2.0)+1.0);
 
 static const unsigned int min_alignment = 1;
 static const unsigned int max_alignment = 8;
